requirements.txt
gradio>=4.31.5
pandas>=2.2.2
nltk>=3.8.1
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0.post1
langchain>=0.2.12
langchain-community>=0.2.12
langchain-text-splitters>=0.2.2
unstructured>=0.15.0
pypdf>=4.2.0
python-docx>=1.1.2
python-pptx>=0.6.23
openpyxl>=3.1.5
PyYAML>=6.0.1
pydantic>=2.7
# interlinked is an internal dependency; ensure it is installed in your environment
